requirements.txt
pytest
mock
sh
PyYAML
python-gnupg
